+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added link on the About page to let you check for a newer version of the app [#18](https://github.com/cheshire137/competiwatch-desktop/issues/18)
 - Added "Export matches" link to the Matches page to allow downloading your match history for that account and season [#32](https://github.com/cheshire137/competiwatch-desktop/issues/32)
 - Added "Export Matches" menu item to the Tools menu when viewing a particular account's match history
+- Fixed bug where the the "New SR" field wouldn't have your latest SR when you used the "Log a Match" menu option/hit Option-L [#10](https://github.com/cheshire137/competiwatch-desktop/issues/10)
 
 ## Version 0.1.6